Add Support for DRF-like ModelSerializer
This was achieved by adding a check if there is a `Meta` class present inside
the custom serializer.
Now, if that class exists, an attempt is made to get the model information for
both Django models as well as SQLAlchemy models. Others are not supported and
will raise a RuntimeError.
The `_field_map` is then updated accordingly, but MethodFields or custom `attr`
keyword arguments are not overwritten.