Fix bug overwriting explicitly declared fields
Before, the `direct_fields` dict holding all fields declared in the custom
serializer was overwritten with all implicitly defined fields from the `Meta`
class inside the custom serializer.

That means if someone declared anything other than a standard Field for an
attribute of the model, this would have been overwritten with a standard Field
instance instead of keeping the explicitly declared field type, effectively
nullifying the effect of the explicit declaration.

Now, if someone was to declare a field explicitly, the SerializerMeta class
will not overwrite this field  with a standard Field instance in the
`direct_fields` dict, but will instead skip this field, thus preserving the
explicit declaration.

Also add some more comments explaining this in the code.